+++ docs/trace/getting-started-aspnetcore/Program.cs
@@ -27,8 +27,7 @@ appBuilder.Services.AddOpenTelemetry()
         .AddService(serviceName: "OTel.NET Getting Started"))
     .WithTracing(builder => builder
         .AddAspNetCoreInstrumentation()
-        .AddConsoleExporter())
-    .StartWithHost();
+        .AddConsoleExporter());
 
 var app = appBuilder.Build();