+++ examples/AspNetCore/Program.cs
@@ -44,8 +44,7 @@ Action<ResourceBuilder> configureResource = r => r.AddService(
 // for manual instrumentation
 appBuilder.Services.AddSingleton<Instrumentation>();
 
-// Configure OpenTelemetry tracing & metrics with auto-start using the
-// StartWithHost extension from OpenTelemetry.Extensions.Hosting.
+// Configure OpenTelemetry tracing & metrics.
 appBuilder.Services.AddOpenTelemetry()
     .ConfigureResource(configureResource)
     .WithTracing(builder =>
@@ -127,8 +126,7 @@ appBuilder.Services.AddOpenTelemetry()
                 builder.AddConsoleExporter();
                 break;
         }
-    })
-    .StartWithHost();
+    });
 
 // Clear default logging providers used by WebApplication host.
 appBuilder.Logging.ClearProviders();