+++ src/OpenTelemetry/Internal/OpenTelemetrySdkEventSource.cs
@@ -175,6 +175,15 @@ namespace OpenTelemetry.Internal
             }
         }
 
+        [NonEvent]
+        public void HostedServiceRegistrationFailure(Exception ex)
+        {
+            if (this.IsEnabled(EventLevel.Error, EventKeywords.All))
+            {
+                this.HostedServiceRegistrationFailure(ex.ToInvariantString());
+            }
+        }
+
         [Event(1, Message = "Span processor queue size reached maximum. Throttling spans.", Level = EventLevel.Warning)]
         public void SpanProcessorQueueIsExhausted()
         {
@@ -413,6 +422,36 @@ namespace OpenTelemetry.Internal
         public void InvalidEnvironmentVariable(string key, string value)
         {
             this.WriteEvent(47, key, value);
+        }
+
+        [Event(48, Message = "OpenTelemetry IHostedService registered in application services.", Level = EventLevel.Informational)]
+        public void HostedServiceRegistered()
+        {
+            this.WriteEvent(48);
+        }
+
+        [Event(49, Message = "OpenTelemetry IHostedService application services registration skipped. Reason: '{0}'", Level = EventLevel.Warning)]
+        public void HostedServiceRegistrationSkipped(string reason)
+        {
+            this.WriteEvent(49, reason);
+        }
+
+        [Event(50, Message = "OpenTelemetry IHostedService could not be registered in application services. Error: '{0}'", Level = EventLevel.Error)]
+        public void HostedServiceRegistrationFailure(string error)
+        {
+            this.WriteEvent(50, error);
+        }
+
+        [Event(51, Message = "OpenTelemetry TracerProvider was not found in application services. Tracing will remain disabled.", Level = EventLevel.Warning)]
+        public void TracerProviderNotRegistered()
+        {
+            this.WriteEvent(51);
+        }
+
+        [Event(52, Message = "OpenTelemetry MeterProvider was not found in application services. Metrics will remain disabled.", Level = EventLevel.Warning)]
+        public void MeterProviderNotRegistered()
+        {
+            this.WriteEvent(52);
         }
 
 #if DEBUG