+++ src/OpenTelemetry/OpenTelemetryServiceCollectionExtensions.cs
@@ -17,6 +17,7 @@
 #nullable enable
 
 using Microsoft.Extensions.DependencyInjection;
+using OpenTelemetry.Internal;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
 
@@ -35,12 +36,12 @@ public static class OpenTelemetryServiceCollectionExtensions
     /// Notes:
     /// <list type="bullet">
     /// <item>A <see cref="TracerProvider"/> and/or <see cref="MeterProvider"/>
-    /// will not be created automatically using this method. To begin collecting
-    /// traces and/or metrics either use the
-    /// <c>OpenTelemetryBuilder.StartWithHost</c> extension in the
-    /// <c>OpenTelemetry.Extensions.Hosting</c> package or access the <see
-    /// cref="TracerProvider"/> and/or <see cref="MeterProvider"/> through the
-    /// application <see cref="IServiceProvider"/>.</item>
+    /// will be created automatically using this method if a host supporting
+    /// <c>Microsoft.Extensions.Hosting.IHostedService</c> is detected at
+    /// runtime. To begin collecting traces and/or metrics when hosting
+    /// extensions are not being used access the <see cref="TracerProvider"/>
+    /// and/or <see cref="MeterProvider"/> through the application <see
+    /// cref="IServiceProvider"/>.</item>
     /// <item>This is safe to be called multiple times and by library authors.
     /// Only a single <see cref="TracerProvider"/> and/or <see
     /// cref="MeterProvider"/> will be created for a given <see
@@ -51,5 +52,9 @@ public static class OpenTelemetryServiceCollectionExtensions
     /// <returns>The supplied <see cref="OpenTelemetryBuilder"/> for chaining
     /// calls.</returns>
     public static OpenTelemetryBuilder AddOpenTelemetry(this IServiceCollection services)
-        => new(services);
+    {
+        HostingHelper.AddOpenTelemetryHostedServiceIntoServiceCollection(services);
+
+        return new(services);
+    }
 }